Architecture total revamp! Need feedback.
