+++ src/goban/fscore.rs
