+++ src/lib.rs
@@ -12,15 +12,13 @@ mod algorithm;
 mod bitboard;
 mod goban;
 mod node;
-mod stone;
-mod fscore;
 
 // Comment rajouter une fonction python sur rust
 // Simplement rajouter dans le block pymodule une fonction rust avec obligatoirement une instance Python<'_>, et si applicable un PyResult pour le retour
 // Presque n'importe quel type peut etre passé tant que c'est un type natif python/rust (check doc)
 // Pour compiler, maturin develop dans le terminal, qui genere un dylib dans le dossier target/debug qu'il faut mettre dans le dossier root du projet.
 
-const DEPTH: u32 = 1;
+const DEPTH: u32 = 9;
 const WHITE: u8 = 2;
 const BLACK: u8 = 1;
